Refactor to simplify request creation
