+++ conjure-go-client/httpclient/request_builder.go
@@ -20,8 +20,6 @@ import (
 	"net/url"
 
 	"github.com/palantir/pkg/bytesbuffers"
-	"github.com/palantir/witchcraft-go-error"
-	"github.com/palantir/witchcraft-go-tracing/wtracing"
 )
 
 type requestBuilder struct {
@@ -47,70 +45,4 @@ type requestParamFunc func(*requestBuilder) error
 
 func (f requestParamFunc) apply(b *requestBuilder) error {
 	return f(b)
-}
-
-// NewRequest returns an *http.Request and a set of Middlewares which should be
-// the inner-most wrapped middleware around the request during execution.
-func (c *clientImpl) newRequest(ctx context.Context, baseURL string, params ...RequestParam) (*http.Request, []Middleware, bool, error) {
-	b := &requestBuilder{
-		headers:        c.initializeRequestHeaders(ctx),
-		query:          make(url.Values),
-		bodyMiddleware: &bodyMiddleware{bufferPool: c.bufferPool},
-	}
-
-	for _, p := range params {
-		if p == nil {
-			continue
-		}
-		if err := p.apply(b); err != nil {
-			return nil, nil, false, err
-		}
-	}
-	for _, c := range b.configureCtx {
-		ctx = c(ctx)
-	}
-
-	if b.method == "" {
-		return nil, nil, false, werror.Error("httpclient: use WithRequestMethod() to specify HTTP method")
-	}
-
-	req, err := http.NewRequest(b.method, baseURL+b.path, nil)
-	if err != nil {
-		return nil, nil, false, werror.Wrap(err, "failed to build new HTTP request")
-	}
-	req = req.WithContext(ctx)
-	req.Header = b.headers
-	if q := b.query.Encode(); q != "" {
-		req.URL.RawQuery = q
-	}
-
-	// the middleware must be wrapped in a certain order, with the inner-most receiving the request last and the
-	// response first.
-	var innerMiddleware []Middleware
-	for _, m := range []Middleware{
-		// must precede the error decoders because they return a nil response and the metrics need the status code of
-		// the raw response.
-		c.metricsMiddleware,
-		// must precede the client error decoder
-		b.errorDecoderMiddleware,
-		// must precede the body middleware so it can read the response body
-		c.errorDecoderMiddleware,
-		b.bodyMiddleware,
-	} {
-		if m != nil {
-			innerMiddleware = append(innerMiddleware, m)
-		}
-	}
-	return req, innerMiddleware, !b.bodyMiddleware.rawOutput, nil
-}
-
-func (c *clientImpl) initializeRequestHeaders(ctx context.Context) http.Header {
-	headers := make(http.Header)
-	if !c.disableTraceHeaderPropagation {
-		traceID := wtracing.TraceIDFromContext(ctx)
-		if traceID != "" {
-			headers.Set(traceIDHeaderKey, string(traceID))
-		}
-	}
-	return headers
 }
